[BUGFIX] Handle submodule includes
While registering a module like defined in README, the variable
`$allowedModule` is set to true in submodules. This leads to a TypeError
while array_merge().

To avoid this, a check is added if the referring submodule is loaded and
then appends the submodule to the current `$finalModules` list.

As `$moduleName` could be assigned to the main module name, the
assignment at the beginning of the foreach was removed.